requirements.txt
requests==2.31.0
python-dotenv==1.0.1
langgraph==0.2.69
langchain==0.3.17
langchain-openai==0.3.3
langchain-core==0.3.33
langchain-community==0.3.*
langchain-chroma==0.2.*
networkx==3.0.0
chromadb==0.6.0
